requirements.lock.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-default-group==1.2.4
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
databricks-sdk==0.91.0
datasets==4.5.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
docker==7.1.0
einops==0.8.2
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.129.0
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
formulaic==1.2.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
interface-meta==1.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
linearmodels==7.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mlflow==3.10.0
mlflow-skinny==3.10.0
mlflow-tracing==3.10.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.16.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook_shim==0.2.4
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
open_clip_torch==3.2.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.1
pandas-flavor==0.8.1
pandocfilters==1.5.1
parso==0.8.6
patsy==1.0.2
peft==0.18.1
pexpect==4.9.0
pillow==12.0.0
pingouin==0.5.5
platformdirs==4.9.2
playwright==1.58.0
pluggy==1.6.0
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pyhdfe==0.2.0
pynvml==13.0.1
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
selenium==4.41.0
Send2Trash==2.1.0
sentence-transformers==5.2.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlite-fts4==1.0.3
sqlite-utils==3.39
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
statsmodels==0.14.6
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2026.2.20
timm==1.0.24
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0+cu128
torchaudio==2.10.0+cu128
torchvision==0.25.0+cu128
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
trio==0.33.0
trio-websocket==0.12.2
triton==3.6.0
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
webcolors==25.10.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
wsproto==1.3.2
xarray==2026.2.0
xformers==0.0.35
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
